requirements.txt
tensorflow==2.17.0
pandas~=2.2.2
opencv-python==4.10.0.84
PyYAML~=6.0.2rc1
pytest~=6.2.5
jsonargparse==4.32.0
scikit-learn~=1.5.0
typing~=3.7.4.3
matplotlib~=3.9.0
numpy~=1.26.4
keras~=3.3.3
setuptools~=75.1.0
seaborn==0.13.1
jupyter==1.0.0
pip==24.2
attrs==23.2.0
wheel==0.43.0
tornado==6.4
filelock==3.15.4